+++ transports/quic/src/lib.rs
@@ -73,6 +73,7 @@ pub use connection::{Connecting, Connection, Stream};
 #[cfg(feature = "tokio")]
 pub use provider::tokio;
 pub use provider::Provider;
+pub use socket2;
 pub use transport::GenTransport;
 
 /// Errors that may happen on the [`GenTransport`] or a single [`Connection`].